Vida para testear
En un principio he terminado, y ya debería funcionar todo correctamente. Avisadme si hay algún fallo que yo no he visto mientras experimentaba con la barra de vida.

+++ Assets/Scripts/Dummy/HarmIndicatorManager.cs
@@ -52,7 +52,7 @@ public class HarmIndicatorManager : MonoBehaviour
         //DamageText.text = $"{damage}";
     }
 
-    public void AskForHealth(int damage)
+    public void AskForHealth(float damage)
     {
         Canvas text = Instantiate<Canvas>(DamageText, gameObject.transform);
         text.GetComponent<DamageNumberScript>().SetText(damage.ToString());
