+++ Assets/Scripts/Enemy/BasicEnemy/EnemyAttackState.cs
@@ -30,7 +30,7 @@ public class EnemyAttackState : BaseState
     /// <summary>
     /// El daño del ataque basico
     /// </summary>
-    [SerializeField] int _damage;
+    [SerializeField] float _damage;
 
     #endregion
